pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["Matheus Correia"]

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"
pytest = "^7.1.2"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.18.1"}
PyJWT = "^2.4.0"
SQLAlchemy = "^1.4.39"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
